Duc/native game/implement native overlay (#768)
* impl native overlay

* UI

* missing

* upgrade google mobile ads version

* update version admob

* Remove commented-out AdMob using statement

Cleaned up unused and commented-out code for better readability and maintainability. This ensures no confusion or redundancy in the codebase.

* update remove ads

* add log

* upgrade admob version

* remove native package

* Add [Preserve] attribute to NativeOverlayInterPopupPresenter

This ensures the method is not stripped during the Unity build process when the code stripping feature is enabled. Also added the UnityEngine.Scripting namespace to support the [Preserve] attribute.

* update

* update native overlay inter

* update counter

* update

* hide mrec when show native overlay

* Hide MREC when showing native overlay ad

Added a condition to ensure the MREC ad is hidden when displaying the native overlay ad. This prevents overlapping ads and maintains a seamless user experience. Updated IntelliJ workspace to reflect recent changes.

* update time, comment

* update symbol